backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
python-decouple
fastapi-mail
